// SDKs/Widget SDKs/Infor_WidgetSDK_3.0.1/samples/infor.sample.angular.contenttranslation/manifest-types.d.ts
// NOTE: This file has been automatically generated. It should never be manually edited.
import { ILanguage } from "@infor-lime/core";
export interface IManifestLanguage extends ILanguage<IManifestLanguage> {
	/** Add */
	add: string;
	/** Description */
	description: string;
	/** Edit item */
	editItem: string;
	/** Title */
	title: string;
	/** Content Translation sample widget. */
	widgetDescription: string;
	/** Content Translation */
	widgetTitle: string;
}
export type IManifestSettings = Record<string, never>;